"Intercomponent communication" (delete contact)
Write code to delete a contact in response to a button press. In the Om
example, the author implemented this feature using channels to implement
communication between components.

In the Reagent example, we do not need channels; we simply invoke a
function that removes the specified contact from the global state.

I'm unclear if the asynchronous communication in the Om example is
needed in a real application. We'll see.